% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licloc.R
\name{lic}
\alias{lic}
\title{Wrapper function for tic}
\usage{
lic(
  ...,
  quiet = FALSE,
  func.tic = NULL,
  msg = NULL,
  namespace = NA_character_,
  .logcall = sys.call(),
  .topcall = sys.call(-1),
  .topenv = parent.frame(),
  level = licloc.get_log_level()
)
}
\arguments{
\item{...}{Compiled into a message if \code{msg} is NULL}

\item{quiet}{Passthru to  \code{\link[tictoc]{tic}}}

\item{func.tic}{Passthru to  \code{\link[tictoc]{tic}}}

\item{msg}{Passthru to  \code{\link[tictoc]{tic}}}

\item{namespace}{Passthru to \code{\link[logger]{log_level}}}

\item{.logcall}{Passthru to \code{\link[logger]{log_level}}}

\item{.topcall}{Passthru to \code{\link[logger]{log_level}}}

\item{.topenv}{Passthru to \code{\link[logger]{log_level}}}

\item{level}{Passthru to \code{\link[logger]{log_level}}}
}
\value{
Nothing
}
\description{
Uses nested messages and, if desired, logging over printing to console
}
