% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licloc.R
\name{licloc.get_indent_string}
\alias{licloc.get_indent_string}
\title{Returns a string for left padding based on the current indentation level}
\usage{
licloc.get_indent_string()
}
\value{
string used for left padding
}
\description{
Returns a string for left padding based on the current indentation level
}
