% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licloc.R
\name{licloc.get_log_level}
\alias{licloc.get_log_level}
\title{Get the logging threshold for licloc}
\usage{
licloc.get_log_level()
}
\value{
the logging threshold for licloc
}
\description{
Get the logging threshold for licloc
}
