% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licloc.R
\name{licloc.get_symbol_at}
\alias{licloc.get_symbol_at}
\title{Gets the symbol for a specific indentation level}
\usage{
licloc.get_symbol_at(num)
}
\description{
Gets the symbol for a specific indentation level
}
