% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licloc.R
\name{licloc.get_symbol_string}
\alias{licloc.get_symbol_string}
\title{Returns a string with the symbol for the current indentation level}
\usage{
licloc.get_symbol_string()
}
\description{
Returns a string with the symbol for the current indentation level
}
