% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licloc.R
\name{licloc.set_log_level}
\alias{licloc.set_log_level}
\title{Sets the level at which tictoc messages should be logged}
\usage{
licloc.set_log_level(level = logger::DEBUG)
}
\arguments{
\item{level}{The logging threshold which tictoc messages should use}
}
\description{
Setting this value means that tictoc will use logger as opposed to printing
to the screen directly
}
